requirements.txt
requests
argparse
ffmpeg-python
